+++ src/main/java/com/chainXpert/fin_manager/service/UserService.java
@@ -49,7 +49,7 @@ public class UserService {
 
     public ResponseEntity<?> retreive(String token) {
         final var userId = jwtUtils.extractUserId(token.replace("Bearer ", ""));
-        final var user = userRepository.findById(userId).get();
+        final var user = userRepository.findById(userId.longValue()).get();
         return ResponseEntity.ok(UserDetailsDto.builder().createdAt(user.getCreatedAt())
                 .dateOfBirth(user.getDateOfBirth()).emailId(user.getEmailId()).firstName(user.getFirstName())
                 .lastName(user.getLastName()).updatedAt(user.getUpdatedAt()).build());
@@ -102,7 +102,7 @@ public class UserService {
     public ResponseEntity<?> updatePassword(final UserPasswordUpdationRequestDto userPasswordUpdationRequestDto,
                                             final String token) {
         final var loggedInUserId = jwtUtils.extractUserId(token.replace("Bearer ", ""));
-        final var user = userRepository.findById(loggedInUserId).get();
+        final var user = userRepository.findById(loggedInUserId.longValue()).get();
 
         if (!passwordEncoder.matches(userPasswordUpdationRequestDto.getOldPassword(), user.getPassword()))
             return responseUtils.invalidPasswordResponse();
@@ -115,14 +115,14 @@ public class UserService {
 
     public ResponseEntity<?> deleteAccount(final String token) {
         final var userId = jwtUtils.extractUserId(token.replace("Bearer ", ""));
-        final var totalBalance = totalBalanceRepository.findByUserId(userId).get();
+        final var totalBalance = totalBalanceRepository.findByUserId(userId.longValue()).get();
         totalBalanceRepository.deleteById(totalBalance.getId());
         return ResponseEntity.ok().build();
     }
 
     public ResponseEntity<?> update(final UserDetailUpdationRequestDto userDetailUpdationRequest, final String token) {
         final var loggedInUserId = jwtUtils.extractUserId(token.replace("Bearer ", ""));
-        final var user = userRepository.findById(loggedInUserId).get();
+        final var user = userRepository.findById(loggedInUserId.longValue()).get();
 
         user.setFirstName(userDetailUpdationRequest.getFirstName());
         user.setLastName(userDetailUpdationRequest.getLastName());
